+++ src/hyperlight_host/src/lib.rs
@@ -89,6 +89,8 @@ pub use hypervisor::virtual_machine::is_hypervisor_present;
 /// A sandbox that can call be used to make multiple calls to guest functions,
 /// and otherwise reused multiple times
 pub use sandbox::MultiUseSandbox;
+/// The lifecycle state of a [`MultiUseSandbox`].
+pub use sandbox::SandboxStatus;
 /// The re-export for the `UninitializedSandbox` type
 pub use sandbox::UninitializedSandbox;
 /// The re-export for the `SandboxBuilder` type